Add bookmarklet support
Berserk will now show a bookmarklet at the bottom of the edit page that can be
used to add a bug to the current sprint from a Bugzilla page.  This _requires_
that the Sites table be properly set up, else the bookmarklet link will be
incorrect.  Also, include django-flash for error message goodness.